:truck: [Utilities] Move isNewDesignLanguageColor to a new file
