+++ src/components/Icon/Icon.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 
-import {useFeatures} from '../../utilities/features';
 import {classNames, variationName} from '../../utilities/css';
+import {isNewDesignLanguageColor} from '../../utilities/color-new-design-language';
+import {useFeatures} from '../../utilities/features';
 import {useI18n} from '../../utilities/i18n';
-import {IconProps, isNewDesignLanguageColor} from '../../types';
+import type {IconProps} from '../../types';
 
 import styles from './Icon.scss';
