+++ src/index.ts
@@ -24,6 +24,8 @@ export {
   rgbaString,
 } from './utilities/color-transformers';
 
+export {isNewDesignLanguageColor} from './utilities/color-new-design-language';
+
 export {ScrollLockManagerContext as _SECRET_INTERNAL_SCROLL_LOCK_MANAGER_CONTEXT} from './utilities/scroll-lock-manager';
 export {WithinContentContext as _SECRET_INTERNAL_WITHIN_CONTENT_CONTEXT} from './utilities/within-content-context';
